edit styles, edit launch point for MoreFilters
