+++ packages/ui/modals/index.ts
@@ -1,3 +1,14 @@
-// codegen:start {preset: barrel, exclude: ./*.stories.*}
+// codegen:start {preset: barrel, include: ./*.ts*, exclude: ./*.stories.tsx}
+export * from './DeleteModal'
+export * from './Example'
+export * from './ForgotPassword'
+export * from './Login'
+export * from './ModalTitle'
+export * from './MoreFilter'
+export * from './PrivacyStatement'
+export * from './QuickPromotion'
+export * from './Review'
+export * from './Service'
+export * from './ServiceFilter'
 export * from './settings'
 // codegen:end
